Kept the interface to keep the API from breaking
